Add a new error message in hardhat-etherscan.
This error only applies when passing a fully qualified name.
It communicates the fact that a different compiler was used to
generate the retrieved bytecode rather than the one Hardhat currently
has configured for this particular contract.

Add a couple of integration tests for the fully qualified name parameter.
Update the error message that communicates the contract detection
ambiguity. Now it explains how it can be solved.