Load undici dynamically
If we don't do this, then undici is part of the modules that are unloaded when `resetHardhatContext` is used. After some repeated usages (around 100), the `hardhat-network-helpers` tests start being noticeably slower. We traced this to undici being unloaded, but we don't really understand the reason. In any ase, just not loading it as part of the context intialization fixes the problem.