Avoid maxExtraDataSize validation when processing remote blocks
When a remote block is processed, it might happen that its `extraData` field is longer than 32 bytes. This throws a validation error if the common being used has PoW as its consensus algorithm (which Hardhat Network has). To prevent this, we try to use the proper common for the remote network (if possible), and just use the one for Goerli if not.