Re-throw errors in vm-debug-tracer.ts
If there was an error in one of the event handlers of this tracer, the tests would just timeout. This happened many times during the EthereumJS upgrade. To make this better, now any error thrown within the step handler (which has the more complex logic) is re-thrown in the trace method.