Enforce leading _ in private class properties
We were already following this convention, but eslint wasn't enforcing it correctly because the @typescript-eslint/naming-convention rule uses the most specific entry in the config. The leading underscores rule was specified with the memberLike selector, but there's also a classProperty selector that is more specific. Adding a new, even more specific entry for private classProperties fixes the problem.