Create common with proper hardfork before manipulating the block
This prevents the following bug:
- We use the merge hardfork locally
- We fork a shanghai block
- We run a call in a block before the forked block

In this case, the remote block will be fetched and will have
withdrawals. But when Block.fromBlockData is called with the current
common (which has a merge hardfork) it will fail. If we create the right
common before doing the block context manipulation, this doesn't happen.

There is already a test for this scenario. I don't know why it wasn't
failing before this.