Hardcode max precompile in VMTracer
Temporarily hardcode the max precompile number in VMTracer to remove the
need of using ethereumjs' common and evm. This is a temporary solution
until we complete the port.

The worst-case scenario with this change is that someone could get a
precompile trace for an address that doesn't actually have a precompile.
For example: someone is using the `shanghai` hardfork, they send a tx to
address `0xA` (which only has a precompile starting from `cancun`), and
they get a precompile trace for that execution. This is seems like a
very unlikely scenario to happen, and it will be a temporary situation
anyway.